+++ src/Apps/Artwork/Components/__tests__/PricingContext.test.tsx
@@ -1,4 +1,4 @@
-import { BarBox, BarChart, QuestionCircleIcon } from "@artsy/palette"
+import { BarBox, BarChart, Link, QuestionCircleIcon } from "@artsy/palette"
 import { mockTracking } from "Artsy/Analytics"
 import { renderRelayTree } from "DevTools"
 import { mount } from "enzyme"
@@ -16,6 +16,7 @@ jest.unmock("react-relay")
 const mockPricingContext = {
   appliedFiltersDisplay: "Price ranges of small mocks by David Sheldrick",
   filterDescription: `deprecated field`,
+  appliedFilters: { category: "PAINTING", dimension: "SMALL" },
   bins: [
     {
       maxPrice: "$88",
@@ -50,8 +51,6 @@ const mockPricingContext = {
 
 const mockArtwork = {
   artists: [{ id: "artist-id" }],
-  widthCm: 234,
-  heightCm: 234,
   category: "Photography",
   pricingContext: mockPricingContext,
   priceCents: {
@@ -277,6 +276,23 @@ Object {
         context_module: "Price Context",
         action_type: "Hover",
         subject: "Histogram Bar",
+        flow: "Artwork Price Context",
+        type: "Chart",
+      })
+      expect(dispatch).toHaveBeenCalledTimes(1)
+    })
+
+    it("tracks clicks on 'Browse works in this category' link", () => {
+      const { Component, dispatch } = mockTracking(PricingContext)
+      const component = mount(<Component artwork={mockArtwork as any} />)
+      component
+        .find(Link)
+        .at(0)
+        .simulate("click")
+      expect(dispatch).toBeCalledWith({
+        context_module: "Price Context",
+        action_type: "Click",
+        subject: "Browse works in this category",
         flow: "Artwork Price Context",
         type: "Chart",
       })