+++ src/Components/v2/__stories__/AuctionCard.story.tsx
@@ -12,6 +12,8 @@ storiesOf("Styleguide/Components", module).add("AuctionCard", () => {
           headline="Sotheby’s"
           subHeadline="Contemporary Day Sale"
           badge="In progress"
+          isGalleryAuction={false}
+          isBenefit={false}
           href="#"
         />
       </Section>
@@ -21,6 +23,19 @@ storiesOf("Styleguide/Components", module).add("AuctionCard", () => {
           headline="Sotheby’s"
           subHeadline="Contemporary Day Sale"
           badge="In progress"
+          isGalleryAuction={false}
+          isBenefit={false}
+          href="#"
+        />
+      </Section>
+      <Section title="Large Auction Card for benefit auction">
+        <LargeAuctionCard
+          src="https://picsum.photos/200/180/?random"
+          headline="MOMA Benefit Auction"
+          subHeadline="MOMA Benefit Auction"
+          badge="In progress"
+          isGalleryAuction={false}
+          isBenefit
           href="#"
         />
       </Section>
@@ -30,6 +45,19 @@ storiesOf("Styleguide/Components", module).add("AuctionCard", () => {
           headline="Sotheby’s"
           subHeadline="Contemporary Day Sale"
           badge="In progress"
+          isGalleryAuction={false}
+          isBenefit={false}
+          href="#"
+        />
+      </Section>
+      <Section title="Small Auction Card for gallery auction">
+        <SmallAuctionCard
+          src="https://picsum.photos/200/180/?random"
+          headline="Pace Gallery"
+          subHeadline="Pace Gallery"
+          badge="In progress"
+          isGalleryAuction
+          isBenefit={false}
           href="#"
         />
       </Section>