+++ src/__generated__/ArtworkBanner_artwork.graphql.ts
@@ -17,7 +17,8 @@ export type ArtworkBanner_artwork = {
     }) | null;
     readonly sale: ({
         readonly is_auction: boolean | null;
-        readonly is_benefit: boolean | null;
+        readonly isBenefit: boolean | null;
+        readonly isGalleryAuction: boolean | null;
         readonly cover_image: ({
             readonly url: string | null;
         }) | null;
@@ -240,7 +241,14 @@ return {
         {
           "kind": "ScalarField",
           "alias": null,
-          "name": "is_benefit",
+          "name": "isBenefit",
+          "args": null,
+          "storageKey": null
+        },
+        {
+          "kind": "ScalarField",
+          "alias": null,
+          "name": "isGalleryAuction",
           "args": null,
           "storageKey": null
         },
@@ -425,5 +433,5 @@ return {
   ]
 };
 })();
-(node as any).hash = '1088a23ad9988055e5834161ce07a18f';
+(node as any).hash = 'a01d8a22bebc8854e61f1d5d99ebdea7';
 export default node;